requirements.txt
langchain==1.2.15
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_core==1.2.28
langchain_text_splitters==1.1.1
langgraph==1.1.6
PyYAML==6.0.3
streamlit==1.56.0
dashscope==1.25.16